Modified players struct: now cards fields are no more pointers to Cards. Modified deck in main: now it is just an array in the heap of Cards. Code adjustments due these modifications.
Defined and implemented functions to announce dead players, remove dead players, count dead players, withdraw the cards at the end of a game phase, give back new cards to the players before a new game phase starts.

Improvements at function handleGamePhase.
Fixed bugs in function applyEffect that caused segmentation faults.
Improvements at function applyEffect: in particular, in previous versions, it assumed that the player id was equal to the index + 1 of the player in the players' array; therefore, applyEffect worked as expected until no player died.

Defined and implemented functions to print all players and to print the deck.
Now printPlayer takes a player by value and not by pointer.

Defined and implemented the function Pause that halts the game until the user presses a key in the keyboard. If clear is set to true, it also clears the screen; this function is exploited in handleGamePhase in order to not make the console too much verbose, now at the end of each player turn it halts and clears the screen.

Main function has been finalized.