Upgrading code to more recent standards.
Now properties in builder functions are referred to as .<property-name> = and not as <property-name>:, which was obselete.
Now boolean properties of structs in builder functions are set to false and not to NONE: in the previous versions of code, if these properties had not been modified anymore, they would have made the function printGraphicRendition behave as the would have been set to true. Modifications to the function printGraphicRendition have been done due this bug.