Various edits.
In h/config.h, the struct gameConfiguration has been moved to file h/game_configuration.h. Adapted the include directives due to this modification. Defined and implemented the function to load the correct game configuration: the function getGameConfiguration checks, with the help of the internal action code, if it has to load the game configuration either from the configuration file or from the CLI or by asking it to the user.

In h/main.h and src/main.c, a function for printing the game configuration have been defined and implemented. The function askConfigurationOptionsViaTerminal, defined in h/config.h, has been implemented in src/main.c because it needed the ANSI module.

In h/config_file.h and src/config_file.c, functions for reading the game configuration from the configuration file, saving the game configuration to the configuration file and for checking if the configuration file exists have been defined and implemented.