Various edits.
In function getHelp, the help messages have been updated and the help messages for --save-to-file and --verbose options have been added. Now the function getHelp does not exit with EXIT_CLI_ILLEGAL status code.

In function handleCLIArguments, white-spaces clean-up. Handling the help options for getting help for --save-to-file and --verbose options. Handling the illegal help options and exiting with EXIT_CLI_ILLEGAL status code if that is the case. Now if the user asks for help legally, the function returns with the action code ACTION_HELP. Now if the users passes some unrecognized options the program exits with EXIT_CLI_ILLEGAL status code. Handling the --save-to-file and --verbose flags.

In function getConfigurationFromArguments fixed a small bug about the loop iteration.