Various edits.
Now the default configuration has the beVerbose field set to false.

In function getGameConfiguration, the program exits with EXIT_SUCCESS exit code if the action code is ACTION_HELP. Now the function handles the action code with bit masking technique. Now the function saves the configuration to a file if it has to.