Fixed a unsafe-memory management in printfgr.
In function printfgr, previously, the char buffer t was a memory buffer in the heap with size equal to the length of the string text: but the string t, which is the string that contains all the given inputs replaced to the standard C format specifiers, has in the majority of cases a length greater than the length of text; thus, at some point in the execution, the buffer t would have caused requested to access to an invalid memory address. Since the final dimension of the buffer t is complex to determine, the best fit was to use a vector.