Code improvements and adjustments.
In h/main.h, the functions setupTerminal, checkTerminalHost and checkTerminalSize has been documented.
In h/main.h and src/main.c, the struct gameConfiguration has been renamed to GameConfiguration. Function Pause has been moved to ANSI module. Functions printCard, printDeck, printPlayer, printPlayers and printGameConfiguration.
Now the struct Game, contains fields for the game configuration and logs configuration.
Now the function applyEffect returns a boolean value: only with the TUI, it returns true if it has printed some messages, otherwise false. If the TUI is not used, it returns false. This functionality is not completely implemented yet. Now the function decrements safely the lifePoints field of the player's structs. When the function is applying the effect of a card with rank equals to seven, it pauses the game using function Pause. Some messages of the function have been improved.

In src/gui.c, some bugs about the page visualizzation due to print statements of the function applyEffect have been fixed. Small improvements in function drawCardsForPage.
Now the function giveCards does not require anymore the game configuration as a parameter.
Now the function tellFacedDownCard takes as parameter the player ID: if the player ID is set to 0, it prints 'your', otherwise 'Player <player ID>'s'.