Code transfer and improvements.
In h/gui.h and src/gui.c, the function printPageData has been moved to logs module.

In h/logs.h and src/logs.c, moved, adapted and documented the function printPageData. Now the function printCard does not ignore anymore the parameter newLine when the logger is not using the console; documentation of printCard has been adapted in order to reflect this change.